Improve memory allocation made by EnrichTags
EnrichTags use to allocate a slice per dogstatsd sample when origin
detection was enabled. This add a huge impact on k8s setup.

Moving the logic to enrich tag later in the pipeline (in the aggregator)
allow us to create a slice of tags per context (which a renewed every
10s).

Now dogstatsd only set the originID from UDS and from the k8s tag and
let the aggregator pull the tags from the tagger when needed.